services/web/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-oauth-toolkit==1.2.0
django-rest==0.0.1
djangorestframework==3.9.3
idna==2.8
oauthlib==3.0.1
pytz==2019.1
requests==2.21.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.24.3
gunicorn
mysqlclient
psycopg2
django-cors-headers
